Factorise getting workspaceId from datasetId
